requirements.txt
# Google Cloud Services & Authentication
google-api-python-client>=2.154.0
google-auth>=2.36.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1

# AI & Machine Learning
google-generativeai>=0.8.3
prophet>=1.2.1

# Data Processing & Analysis
pandas>=2.2.3
numpy>=2.1.3

# Visualization
plotly>=5.24.1

# Web Framework
streamlit>=1.40.2

# Audio Recording
streamlit-mic-recorder>=0.0.8

# Financial Data
yfinance>=0.2.66

# Date & Time Utilities
python-dateutil>=2.9.0.post0

# Utilities
python-dotenv>=1.0.1
requests>=2.32.3
rich>=13.9.4